Sample dataset from a course about analysis of factorial designs

A sample data set from a course about the analysis of factorial designs, by Mattan S. Ben-Shachar. See following link for more information: https://github.com/mattansb/Analysis-of-Factorial-Designs-foR-Psychologists

The data consists of five variables from 120 observations:

ID: A unique identifier for each participant

sex: The participant's sex

time: The time of day the participant was tested (morning, noon, or afternoon)

coffee: Group indicator, whether participant drank coffee or not ("coffee" or "control").

alertness: The participant's alertness score.

Examples

# Attach coffee-data
data(coffee_data)