Collapse raw data by random effect groups

This function extracts the raw data points (i.e. the data that was used to fit the model) and "averages" (i.e. "collapses") the response variable over the levels of the grouping factor given in collapse_by. Only works with mixed models.

Usage

collapse_by_group(grid, model, collapse_by = NULL, residuals = FALSE)

Arguments

grid

A data frame representing the data grid, or an object of class ggeffects, as returned by predict_response().

model

The model for which to compute partial residuals. The data grid grid should match to predictors in the model.

collapse_by

Name of the (random effects) grouping factor. Data is collapsed by the levels of this factor.

residuals

Logical, if TRUE, collapsed partial residuals instead of raw data by the levels of the grouping factor.

Value

A data frame with raw data points, averaged over the levels of the given grouping factor from the random effects. The group level of the random effect is saved in the column "random".

Examples

library(ggeffects)
data(efc, package = "ggeffects")
efc$e15relat <- as.factor(efc$e15relat)
efc$c161sex <- as.factor(efc$c161sex)
levels(efc$c161sex) <- c("male", "female")
model <- lme4::lmer(neg_c_7 ~ c161sex + (1 | e15relat), data = efc)
me <- predict_response(model, terms = "c161sex")
head(attributes(me)$rawdata)
#>   response x group facet rowname
#> 1       12 2     1     1       1
#> 2       20 2     1     1       2
#> 3       11 1     1     1       3
#> 4       10 1     1     1       4
#> 5       12 2     1     1       5
#> 6       19 1     1     1       6
collapse_by_group(me, model, "e15relat")
#> Collapsing usually not informative across a continuous variable.
#>     x group_col facet rowname random response
#> 1   1         1     1     101      1       14
#> 2   2         1     1     102      1       12
#> 3   2         1     1     103      1       10
#> 4   2         1     1     106      1       14
#> 5   2         1     1     107      1       10
#> 6   1         1     1     108      1       10
#> 7   2         1     1      11      1       28
#> 8   2         1     1     111      1       12
#> 9   1         1     1     114      1        7
#> 10  2         1     1     118      1       16
#> 11  1         1     1     119      1       15
#> 12  2         1     1     120      1       15
#> 13  1         1     1     122      1       17
#> 14  2         1     1     123      1       17
#> 15  2         1     1     124      1       22
#> 16  1         1     1     126      1       14
#> 17  1         1     1     130      1       17
#> 18  2         1     1     131      1       17
#> 19  2         1     1     133      1       13
#> 20  1         1     1     139      1       10
#> 21  1         1     1      14      1       18
#> 22  2         1     1     140      1       12
#> 23  1         1     1     146      1       10
#> 24  2         1     1     148      1       12
#> 25  2         1     1     149      1       11
#> 26  2         1     1     155      1       12
#> 27  1         1     1     156      1       10
#> 28  1         1     1     159      1       12
#> 29  1         1     1     166      1       14
#> 30  2         1     1      17      1       11
#> 31  1         1     1     171      1        9
#> 32  2         1     1     173      1        8
#> 33  2         1     1     174      1       10
#> 34  2         1     1      18      1       11
#> 35  1         1     1     180      1       10
#> 36  2         1     1     183      1       15
#> 37  1         1     1     190      1       12
#> 38  2         1     1     191      1       17
#> 39  1         1     1     192      1       14
#> 40  2         1     1     194      1       12
#> 41  2         1     1     196      1        7
#> 42  2         1     1      20      1       17
#> 43  2         1     1     209      1       12
#> 44  2         1     1     215      1       16
#> 45  2         1     1      22      1        9
#> 46  2         1     1     233      1       10
#> 47  2         1     1     237      1       11
#> 48  2         1     1     253      1       14
#> 49  2         1     1     258      1       15
#> 50  2         1     1     259      1       11
#> 51  2         1     1      26      1       23
#> 52  2         1     1      28      1       11
#> 53  1         1     1       3      1       11
#> 54  1         1     1     302      1        9
#> 55  1         1     1     308      1        9
#> 56  2         1     1     309      1        9
#> 57  2         1     1     310      1        8
#> 58  2         1     1     315      1       12
#> 59  2         1     1     316      1       17
#> 60  2         1     1     318      1       14
#> 61  2         1     1      32      1        9
#> 62  2         1     1      34      1       20
#> 63  1         1     1      35      1        9
#> 64  2         1     1      36      1       10
#> 65  1         1     1     363      1        7
#> 66  2         1     1      37      1       19
#> 67  2         1     1     387      1        7
#> 68  2         1     1      39      1       17
#> 69  2         1     1     396      1       21
#> 70  2         1     1     399      1       15
#> 71  1         1     1       4      1       10
#> 72  1         1     1     400      1        9
#> 73  2         1     1     402      1       12
#> 74  2         1     1     413      1       10
#> 75  1         1     1     414      1       12
#> 76  2         1     1     418      1       11
#> 77  2         1     1      42      1       14
#> 78  2         1     1     423      1       16
#> 79  2         1     1     426      1       14
#> 80  2         1     1      43      1       14
#> 81  1         1     1     435      1       13
#> 82  2         1     1     439      1        9
#> 83  2         1     1      44      1       16
#> 84  1         1     1     448      1       12
#> 85  2         1     1      45      1       19
#> 86  2         1     1     450      1       15
#> 87  2         1     1     462      1       11
#> 88  1         1     1     465      1       13
#> 89  1         1     1      47      1       17
#> 90  1         1     1     477      1       12
#> 91  2         1     1      48      1       15
#> 92  2         1     1     481      1       20
#> 93  1         1     1     483      1        8
#> 94  2         1     1      49      1       16
#> 95  2         1     1     503      1        9
#> 96  1         1     1     508      1       18
#> 97  2         1     1      51      1       17
#> 98  2         1     1     516      1       22
#> 99  1         1     1     518      1       20
#> 100 1         1     1     519      1       19
#> 101 2         1     1      52      1       10
#> 102 2         1     1      53      1       14
#> 103 2         1     1      54      1       14
#> 104 2         1     1     543      1        9
#> 105 2         1     1     547      1        9
#> 106 1         1     1      55      1        9
#> 107 2         1     1     563      1       15
#> 108 1         1     1     568      1        8
#> 109 2         1     1     569      1       13
#> 110 2         1     1      57      1       25
#> 111 2         1     1     572      1       14
#> 112 2         1     1     578      1       15
#> 113 1         1     1     580      1       14
#> 114 2         1     1     584      1       13
#> 115 2         1     1      59      1       17
#> 116 2         1     1     590      1        7
#> 117 2         1     1     593      1       13
#> 118 1         1     1     595      1        7
#> 119 2         1     1      60      1       22
#> 120 2         1     1     613      1       14
#> 121 2         1     1      62      1       15
#> 122 2         1     1     632      1       11
#> 123 2         1     1     634      1       13
#> 124 2         1     1     640      1       13
#> 125 2         1     1     647      1       10
#> 126 2         1     1     649      1       11
#> 127 1         1     1     671      1       11
#> 128 2         1     1     674      1        9
#> 129 1         1     1     676      1       10
#> 130 1         1     1     677      1        9
#> 131 2         1     1     688      1       10
#> 132 1         1     1      69      1       14
#> 133 2         1     1     697      1       23
#> 134 2         1     1       7      1       15
#> 135 1         1     1      70      1       17
#> 136 2         1     1     703      1       18
#> 137 2         1     1     709      1       10
#> 138 2         1     1     718      1        8
#> 139 2         1     1     727      1       13
#> 140 2         1     1     737      1       10
#> 141 2         1     1      74      1       19
#> 142 2         1     1     745      1       12
#> 143 1         1     1     751      1        7
#> 144 2         1     1      76      1       10
#> 145 2         1     1     767      1       13
#> 146 2         1     1     769      1        8
#> 147 2         1     1     780      1       11
#> 148 2         1     1     793      1        9
#> 149 2         1     1      80      1       15
#> 150 2         1     1     804      1       15
#> 151 2         1     1      81      1        7
#> 152 2         1     1     810      1        9
#> 153 1         1     1      82      1       13
#> 154 2         1     1      83      1       12
#> 155 2         1     1     832      1       10
#> 156 2         1     1      87      1       15
#> 157 1         1     1     879      1       24
#> 158 2         1     1      88      1       15
#> 159 2         1     1     886      1       22
#> 160 2         1     1     889      1       11
#> 161 2         1     1     900      1       14
#> 162 2         1     1      91      1       10
#> 163 2         1     1      92      1       11
#> 164 1         1     1      93      1       14
#> 165 2         1     1      95      1       10
#> 166 2         1     1      97      1       12
#> 167 2         1     1      98      1       10
#> 168 2         1     1      99      1       11
#> 169 2         1     1       1      2       12
#> 170 2         1     1      10      2       10
#> 171 2         1     1     104      2        9
#> 172 2         1     1     105      2       13
#> 173 2         1     1     109      2       13
#> 174 2         1     1     110      2       13
#> 175 2         1     1     112      2       12
#> 176 2         1     1     113      2       18
#> 177 2         1     1     115      2       13
#> 178 2         1     1     116      2       14
#> 179 2         1     1     117      2       11
#> 180 1         1     1     121      2        9
#> 181 2         1     1     125      2       16
#> 182 2         1     1     127      2        9
#> 183 2         1     1     128      2       13
#> 184 2         1     1     129      2        9
#> 185 2         1     1      13      2       13
#> 186 2         1     1     132      2        9
#> 187 1         1     1     135      2       14
#> 188 1         1     1     136      2       12
#> 189 2         1     1     137      2       18
#> 190 2         1     1     143      2       14
#> 191 2         1     1      15      2       16
#> 192 2         1     1     150      2       14
#> 193 2         1     1     152      2       13
#> 194 2         1     1     154      2       12
#> 195 2         1     1     157      2        9
#> 196 2         1     1     158      2       15
#> 197 2         1     1      16      2       13
#> 198 2         1     1     162      2       16
#> 199 2         1     1     163      2       18
#> 200 2         1     1     168      2       14
#> 201 2         1     1     169      2       12
#> 202 1         1     1     170      2       10
#> 203 2         1     1     172      2       12
#> 204 2         1     1     175      2       11
#> 205 2         1     1     176      2       10
#> 206 2         1     1     177      2       11
#> 207 2         1     1     178      2       13
#> 208 2         1     1     179      2        7
#> 209 2         1     1     181      2       11
#> 210 2         1     1     185      2       16
#> 211 2         1     1     186      2       13
#> 212 1         1     1     188      2       20
#> 213 1         1     1     189      2        7
#> 214 1         1     1      19      2       13
#> 215 1         1     1     193      2       10
#> 216 2         1     1     195      2        9
#> 217 2         1     1     197      2       16
#> 218 2         1     1     198      2       13
#> 219 2         1     1     199      2       14
#> 220 2         1     1       2      2       20
#> 221 2         1     1     203      2        7
#> 222 2         1     1     204      2       15
#> 223 2         1     1     205      2        7
#> 224 2         1     1     206      2        9
#> 225 2         1     1     207      2       14
#> 226 2         1     1     208      2       11
#> 227 2         1     1      21      2       11
#> 228 2         1     1     210      2       11
#> 229 2         1     1     211      2       12
#> 230 2         1     1     212      2       18
#> 231 2         1     1     213      2        8
#> 232 2         1     1     214      2       13
#> 233 1         1     1     216      2        8
#> 234 2         1     1     217      2       13
#> 235 1         1     1     219      2       11
#> 236 1         1     1     221      2       12
#> 237 2         1     1     222      2       24
#> 238 2         1     1     223      2       11
#> 239 2         1     1     224      2       11
#> 240 2         1     1     225      2       13
#> 241 2         1     1     226      2        9
#> 242 1         1     1     227      2       13
#> 243 2         1     1      23      2        8
#> 244 2         1     1     230      2       16
#> 245 1         1     1     231      2       15
#> 246 1         1     1     232      2       20
#> 247 1         1     1     234      2       12
#> 248 2         1     1     235      2       12
#> 249 1         1     1     236      2       12
#> 250 2         1     1     238      2       10
#> 251 2         1     1     239      2        9
#> 252 2         1     1      24      2       14
#> 253 1         1     1     243      2       12
#> 254 2         1     1     244      2       10
#> 255 2         1     1     246      2       11
#> 256 2         1     1     249      2       11
#> 257 2         1     1      25      2       11
#> 258 2         1     1     250      2       11
#> 259 2         1     1     252      2       11
#> 260 2         1     1     255      2       12
#> 261 2         1     1     256      2       19
#> 262 2         1     1     257      2       12
#> 263 1         1     1     260      2       17
#> 264 2         1     1     261      2        8
#> 265 2         1     1     262      2       13
#> 266 1         1     1     263      2       11
#> 267 2         1     1     264      2       10
#> 268 2         1     1     265      2        8
#> 269 2         1     1     266      2       10
#> 270 2         1     1     267      2       19
#> 271 1         1     1     269      2       10
#> 272 1         1     1     270      2       18
#> 273 2         1     1     271      2        8
#> 274 1         1     1     272      2       11
#> 275 2         1     1     273      2        9
#> 276 1         1     1     274      2       10
#> 277 2         1     1     275      2       13
#> 278 1         1     1     276      2       11
#> 279 1         1     1     277      2        9
#> 280 2         1     1     281      2        9
#> 281 2         1     1     283      2        8
#> 282 2         1     1     286      2       10
#> 283 2         1     1     287      2       12
#> 284 1         1     1     288      2        9
#> 285 2         1     1     289      2       17
#> 286 2         1     1      29      2       15
#> 287 1         1     1     290      2       20
#> 288 2         1     1     291      2       12
#> 289 2         1     1     293      2        9
#> 290 2         1     1     294      2        7
#> 291 2         1     1     296      2        8
#> 292 1         1     1     297      2       14
#> 293 2         1     1      30      2       11
#> 294 2         1     1     301      2       16
#> 295 2         1     1     303      2       16
#> 296 2         1     1     307      2       16
#> 297 1         1     1      31      2       25
#> 298 2         1     1     311      2       15
#> 299 1         1     1     312      2       16
#> 300 1         1     1     313      2       19
#> 301 1         1     1     314      2        8
#> 302 1         1     1     319      2       11
#> 303 2         1     1     320      2        9
#> 304 1         1     1     323      2        8
#> 305 1         1     1     324      2       10
#> 306 1         1     1     325      2        8
#> 307 1         1     1     326      2       10
#> 308 2         1     1     327      2       15
#> 309 2         1     1     328      2       13
#> 310 1         1     1     331      2       10
#> 311 2         1     1     334      2       12
#> 312 1         1     1     336      2        8
#> 313 1         1     1     338      2        9
#> 314 2         1     1     339      2       16
#> 315 2         1     1     341      2        8
#> 316 2         1     1     342      2       13
#> 317 2         1     1     343      2        9
#> 318 2         1     1     344      2        7
#> 319 1         1     1     347      2        8
#> 320 2         1     1     349      2        7
#> 321 2         1     1     351      2        9
#> 322 2         1     1     352      2        7
#> 323 2         1     1     353      2       10
#> 324 1         1     1     354      2       10
#> 325 2         1     1     355      2       11
#> 326 2         1     1     356      2       18
#> 327 1         1     1     357      2        9
#> 328 2         1     1     360      2        8
#> 329 2         1     1     361      2       12
#> 330 2         1     1     362      2        9
#> 331 1         1     1     364      2       10
#> 332 1         1     1     366      2        7
#> 333 2         1     1     369      2       12
#> 334 1         1     1     372      2       10
#> 335 2         1     1     374      2       16
#> 336 2         1     1     377      2       10
#> 337 2         1     1     379      2        9
#> 338 2         1     1      38      2        8
#> 339 2         1     1     380      2       12
#> 340 1         1     1     381      2       18
#> 341 1         1     1     382      2       12
#> 342 2         1     1     384      2       17
#> 343 1         1     1     385      2        9
#> 344 2         1     1     389      2       12
#> 345 2         1     1     390      2       14
#> 346 2         1     1     391      2       15
#> 347 1         1     1     392      2       10
#> 348 1         1     1     394      2       17
#> 349 2         1     1     395      2       17
#> 350 1         1     1     397      2       17
#> 351 2         1     1     398      2       17
#> 352 2         1     1      40      2       16
#> 353 2         1     1     401      2       12
#> 354 2         1     1     403      2       22
#> 355 2         1     1     404      2       11
#> 356 2         1     1     405      2       14
#> 357 2         1     1     406      2       11
#> 358 2         1     1     407      2        9
#> 359 2         1     1     408      2        8
#> 360 2         1     1     409      2       12
#> 361 2         1     1      41      2       17
#> 362 2         1     1     410      2       13
#> 363 2         1     1     411      2       13
#> 364 2         1     1     419      2       17
#> 365 2         1     1     421      2       18
#> 366 2         1     1     422      2       10
#> 367 2         1     1     425      2       10
#> 368 2         1     1     428      2       11
#> 369 2         1     1     430      2       11
#> 370 2         1     1     432      2       15
#> 371 2         1     1     433      2       12
#> 372 2         1     1     434      2       10
#> 373 2         1     1     436      2       13
#> 374 1         1     1     438      2       12
#> 375 2         1     1     440      2       13
#> 376 2         1     1     441      2       10
#> 377 2         1     1     442      2       15
#> 378 2         1     1     444      2       11
#> 379 2         1     1     447      2       12
#> 380 2         1     1     449      2       14
#> 381 2         1     1     451      2        8
#> 382 2         1     1     452      2       10
#> 383 1         1     1     453      2        9
#> 384 2         1     1     456      2       18
#> 385 2         1     1     457      2        7
#> 386 2         1     1     458      2        7
#> 387 2         1     1     460      2       11
#> 388 2         1     1     463      2       11
#> 389 2         1     1     464      2       16
#> 390 2         1     1     466      2       14
#> 391 1         1     1     468      2        7
#> 392 2         1     1     469      2        9
#> 393 2         1     1     471      2       17
#> 394 1         1     1     472      2        7
#> 395 1         1     1     473      2       10
#> 396 2         1     1     474      2        9
#> 397 2         1     1     478      2       14
#> 398 2         1     1     482      2        7
#> 399 2         1     1     485      2       11
#> 400 2         1     1     486      2       10
#> 401 2         1     1     487      2       14
#> 402 1         1     1     488      2        7
#> 403 2         1     1     489      2        8
#> 404 2         1     1     490      2        8
#> 405 2         1     1     493      2       12
#> 406 2         1     1     495      2       14
#> 407 1         1     1     497      2        8
#> 408 1         1     1     498      2       11
#> 409 1         1     1     499      2       17
#> 410 2         1     1       5      2       12
#> 411 2         1     1     500      2       10
#> 412 2         1     1     501      2       22
#> 413 2         1     1     502      2        8
#> 414 1         1     1     505      2       11
#> 415 2         1     1     506      2       18
#> 416 1         1     1     507      2       16
#> 417 1         1     1     509      2       15
#> 418 2         1     1     510      2       19
#> 419 2         1     1     513      2       15
#> 420 2         1     1     514      2        7
#> 421 1         1     1     515      2        8
#> 422 1         1     1     517      2        8
#> 423 1         1     1     520      2       10
#> 424 2         1     1     525      2        7
#> 425 2         1     1     526      2       11
#> 426 2         1     1     528      2        8
#> 427 2         1     1     529      2       12
#> 428 2         1     1     530      2        9
#> 429 2         1     1     532      2        7
#> 430 2         1     1     533      2       12
#> 431 2         1     1     535      2        9
#> 432 2         1     1     536      2        8
#> 433 1         1     1     537      2       10
#> 434 2         1     1     540      2       12
#> 435 2         1     1     545      2        7
#> 436 2         1     1     546      2        8
#> 437 2         1     1     548      2       17
#> 438 2         1     1     549      2       17
#> 439 2         1     1     551      2        9
#> 440 1         1     1     552      2        9
#> 441 2         1     1     554      2        9
#> 442 1         1     1     556      2        8
#> 443 2         1     1     559      2       14
#> 444 2         1     1      56      2       12
#> 445 2         1     1     560      2       28
#> 446 1         1     1     562      2       11
#> 447 2         1     1     564      2        7
#> 448 2         1     1     565      2        9
#> 449 1         1     1     566      2        7
#> 450 1         1     1     567      2        7
#> 451 1         1     1     570      2       19
#> 452 1         1     1     571      2       15
#> 453 2         1     1     573      2       20
#> 454 2         1     1     574      2       20
#> 455 2         1     1     575      2       14
#> 456 2         1     1     576      2       10
#> 457 2         1     1     577      2       11
#> 458 2         1     1     581      2       11
#> 459 2         1     1     583      2       16
#> 460 2         1     1     586      2       17
#> 461 1         1     1     587      2       10
#> 462 2         1     1     589      2       11
#> 463 2         1     1     591      2        8
#> 464 2         1     1     592      2       15
#> 465 2         1     1     594      2       11
#> 466 1         1     1     596      2       18
#> 467 2         1     1     597      2       17
#> 468 2         1     1     598      2       12
#> 469 2         1     1     599      2       18
#> 470 1         1     1       6      2       19
#> 471 1         1     1     603      2       19
#> 472 2         1     1     604      2        7
#> 473 1         1     1     606      2       10
#> 474 2         1     1     608      2       17
#> 475 2         1     1     609      2       19
#> 476 2         1     1      61      2       13
#> 477 1         1     1     610      2        8
#> 478 2         1     1     611      2       12
#> 479 2         1     1     612      2       10
#> 480 2         1     1     614      2       10
#> 481 2         1     1     615      2       13
#> 482 1         1     1     623      2        8
#> 483 2         1     1     625      2       21
#> 484 2         1     1     627      2       11
#> 485 1         1     1     628      2       10
#> 486 2         1     1      63      2        7
#> 487 2         1     1     630      2       11
#> 488 2         1     1     631      2       10
#> 489 2         1     1     635      2       17
#> 490 2         1     1     636      2        9
#> 491 1         1     1     638      2        8
#> 492 1         1     1     639      2        9
#> 493 1         1     1      64      2       19
#> 494 2         1     1     643      2        9
#> 495 2         1     1     644      2       12
#> 496 2         1     1     645      2        8
#> 497 2         1     1     646      2       11
#> 498 2         1     1     648      2       11
#> 499 1         1     1      65      2       11
#> 500 2         1     1     651      2       10
#> 501 2         1     1     652      2       10
#> 502 2         1     1     654      2       13
#> 503 2         1     1     659      2        8
#> 504 2         1     1      66      2       15
#> 505 2         1     1     660      2       13
#> 506 2         1     1     662      2       12
#> 507 2         1     1     663      2       16
#> 508 2         1     1     665      2       10
#> 509 1         1     1     666      2       13
#> 510 1         1     1     667      2       10
#> 511 1         1     1     668      2        9
#> 512 1         1     1     669      2       10
#> 513 2         1     1      67      2       13
#> 514 1         1     1     672      2        9
#> 515 1         1     1     673      2       10
#> 516 2         1     1     678      2       10
#> 517 2         1     1     679      2        8
#> 518 2         1     1     680      2        7
#> 519 2         1     1     683      2        7
#> 520 2         1     1     685      2        8
#> 521 1         1     1     687      2        9
#> 522 2         1     1     690      2       11
#> 523 2         1     1     692      2       16
#> 524 2         1     1     693      2       12
#> 525 1         1     1     694      2       10
#> 526 1         1     1     698      2       10
#> 527 2         1     1     701      2       13
#> 528 2         1     1     702      2       12
#> 529 2         1     1     704      2        9
#> 530 2         1     1     705      2       12
#> 531 2         1     1     706      2       13
#> 532 2         1     1     715      2       11
#> 533 2         1     1     716      2       23
#> 534 2         1     1     719      2        7
#> 535 2         1     1      72      2       15
#> 536 2         1     1     720      2       15
#> 537 2         1     1     722      2       12
#> 538 1         1     1     723      2        9
#> 539 2         1     1     724      2       15
#> 540 2         1     1     725      2       17
#> 541 2         1     1     726      2       13
#> 542 2         1     1     729      2       20
#> 543 2         1     1      73      2       11
#> 544 2         1     1     731      2       11
#> 545 2         1     1     732      2       25
#> 546 2         1     1     734      2       12
#> 547 2         1     1     736      2       12
#> 548 2         1     1     738      2        8
#> 549 2         1     1     739      2       14
#> 550 2         1     1     740      2        8
#> 551 2         1     1     742      2        8
#> 552 2         1     1     743      2       15
#> 553 2         1     1     744      2       11
#> 554 2         1     1     746      2       10
#> 555 2         1     1     747      2        7
#> 556 2         1     1     748      2       10
#> 557 2         1     1      75      2       10
#> 558 1         1     1     750      2       14
#> 559 1         1     1     752      2        7
#> 560 2         1     1     753      2       14
#> 561 2         1     1     754      2       11
#> 562 2         1     1     755      2       11
#> 563 2         1     1     757      2        9
#> 564 2         1     1     759      2       15
#> 565 2         1     1     762      2       18
#> 566 2         1     1     764      2       15
#> 567 2         1     1      77      2       20
#> 568 2         1     1     770      2        9
#> 569 2         1     1     772      2        7
#> 570 2         1     1     776      2       13
#> 571 2         1     1     777      2       10
#> 572 2         1     1     778      2       11
#> 573 1         1     1     779      2       13
#> 574 2         1     1      78      2       10
#> 575 1         1     1     781      2        8
#> 576 2         1     1     782      2       12
#> 577 2         1     1     785      2       16
#> 578 2         1     1     788      2       12
#> 579 2         1     1      79      2       12
#> 580 2         1     1     794      2       13
#> 581 1         1     1     798      2       10
#> 582 2         1     1     799      2       10
#> 583 2         1     1     814      2       13
#> 584 2         1     1     818      2       12
#> 585 1         1     1     819      2        8
#> 586 1         1     1     821      2        7
#> 587 1         1     1     822      2       22
#> 588 2         1     1     823      2       18
#> 589 2         1     1     825      2       13
#> 590 1         1     1     826      2       15
#> 591 1         1     1     827      2       24
#> 592 1         1     1     828      2       11
#> 593 2         1     1     829      2       14
#> 594 2         1     1     830      2       12
#> 595 2         1     1     833      2        7
#> 596 2         1     1     836      2       12
#> 597 2         1     1     838      2        7
#> 598 2         1     1      84      2       16
#> 599 2         1     1     842      2       12
#> 600 2         1     1     843      2        9
#> 601 2         1     1     844      2       10
#> 602 2         1     1     845      2        8
#> 603 1         1     1     846      2        9
#> 604 2         1     1     847      2        7
#> 605 2         1     1     848      2        8
#> 606 2         1     1     849      2       10
#> 607 2         1     1      85      2       10
#> 608 2         1     1     859      2       14
#> 609 2         1     1      86      2       15
#> 610 2         1     1     862      2       10
#> 611 2         1     1     863      2       10
#> 612 1         1     1     864      2       11
#> 613 2         1     1     875      2        7
#> 614 2         1     1     876      2       13
#> 615 2         1     1     877      2       15
#> 616 2         1     1     878      2       11
#> 617 2         1     1     880      2        8
#> 618 2         1     1     881      2        9
#> 619 2         1     1     882      2        7
#> 620 2         1     1     883      2       10
#> 621 2         1     1     885      2       18
#> 622 2         1     1     888      2        9
#> 623 2         1     1     890      2       14
#> 624 2         1     1     891      2        7
#> 625 2         1     1     892      2        9
#> 626 2         1     1     893      2       17
#> 627 2         1     1     894      2       23
#> 628 2         1     1     895      2       12
#> 629 2         1     1     897      2       15
#> 630 1         1     1     899      2        8
#> 631 2         1     1       9      2       15
#> 632 2         1     1      90      2       11
#> 633 2         1     1     901      2       10
#> 634 2         1     1     902      2       10
#> 635 2         1     1     160      3       11
#> 636 1         1     1     229      3       12
#> 637 2         1     1     280      3        8
#> 638 2         1     1     284      3       11
#> 639 2         1     1     299      3       10
#> 640 2         1     1     359      3       13
#> 641 2         1     1     368      3       12
#> 642 1         1     1     416      3       11
#> 643 1         1     1     511      3       10
#> 644 2         1     1     521      3        7
#> 645 2         1     1     544      3        8
#> 646 2         1     1     601      3       13
#> 647 2         1     1     642      3       14
#> 648 1         1     1     686      3       11
#> 649 2         1     1     689      3       12
#> 650 2         1     1     708      3        7
#> 651 2         1     1     712      3       17
#> 652 2         1     1     714      3       14
#> 653 2         1     1     735      3       10
#> 654 1         1     1     756      3       11
#> 655 2         1     1     760      3        9
#> 656 2         1     1     784      3        9
#> 657 2         1     1     792      3       10
#> 658 2         1     1     805      3        9
#> 659 2         1     1     806      3       10
#> 660 2         1     1     807      3        8
#> 661 2         1     1     837      3       10
#> 662 1         1     1     887      3       18
#> 663 1         1     1     138      4        7
#> 664 2         1     1     151      4       11
#> 665 2         1     1     161      4       14
#> 666 2         1     1     164      4       11
#> 667 2         1     1     182      4       12
#> 668 2         1     1     202      4       20
#> 669 2         1     1     218      4       14
#> 670 2         1     1     228      4       20
#> 671 2         1     1     248      4        9
#> 672 2         1     1     295      4        8
#> 673 2         1     1     300      4        8
#> 674 2         1     1     305      4        8
#> 675 2         1     1      33      4       15
#> 676 2         1     1     330      4       10
#> 677 2         1     1     332      4       12
#> 678 2         1     1     337      4       16
#> 679 2         1     1     340      4        7
#> 680 2         1     1     345      4        9
#> 681 1         1     1     348      4       17
#> 682 2         1     1     365      4        8
#> 683 2         1     1     371      4        8
#> 684 2         1     1     373      4       18
#> 685 2         1     1     375      4       11
#> 686 2         1     1     376      4       15
#> 687 1         1     1     378      4        9
#> 688 2         1     1     386      4        8
#> 689 2         1     1     388      4       16
#> 690 2         1     1     420      4       10
#> 691 2         1     1     437      4       13
#> 692 2         1     1     445      4       14
#> 693 1         1     1     461      4        8
#> 694 2         1     1     467      4       14
#> 695 2         1     1     504      4       19
#> 696 2         1     1     512      4       13
#> 697 2         1     1     522      4       25
#> 698 2         1     1     527      4        9
#> 699 2         1     1     531      4       20
#> 700 2         1     1     538      4        8
#> 701 2         1     1     553      4       10
#> 702 2         1     1     555      4        7
#> 703 1         1     1     561      4       16
#> 704 2         1     1     582      4       13
#> 705 2         1     1     588      4       12
#> 706 2         1     1     602      4       10
#> 707 2         1     1     607      4       18
#> 708 1         1     1     618      4        8
#> 709 2         1     1     621      4       11
#> 710 2         1     1     622      4        7
#> 711 2         1     1     626      4        8
#> 712 2         1     1     633      4        9
#> 713 2         1     1     656      4        8
#> 714 2         1     1     657      4       12
#> 715 1         1     1     658      4        8
#> 716 2         1     1     675      4       13
#> 717 1         1     1      68      4       11
#> 718 2         1     1     684      4        9
#> 719 2         1     1     691      4        7
#> 720 2         1     1     696      4       12
#> 721 1         1     1     707      4        9
#> 722 2         1     1     713      4       11
#> 723 2         1     1     717      4       14
#> 724 2         1     1     728      4       10
#> 725 2         1     1     733      4       10
#> 726 2         1     1     741      4       18
#> 727 2         1     1     749      4       13
#> 728 2         1     1     768      4        8
#> 729 1         1     1     771      4        7
#> 730 2         1     1     773      4        9
#> 731 2         1     1     783      4        8
#> 732 2         1     1     786      4       11
#> 733 1         1     1     787      4       19
#> 734 2         1     1       8      4       11
#> 735 2         1     1     803      4       15
#> 736 2         1     1     813      4        9
#> 737 2         1     1     816      4       12
#> 738 2         1     1     834      4       10
#> 739 2         1     1     851      4        8
#> 740 2         1     1     852      4       10
#> 741 2         1     1     854      4        7
#> 742 2         1     1     855      4       11
#> 743 2         1     1     869      4       11
#> 744 2         1     1     884      4       15
#> 745 2         1     1     896      4       13
#> 746 1         1     1      94      4       10
#> 747 2         1     1      96      4       13
#> 748 2         1     1     251      5       12
#> 749 2         1     1     278      5        9
#> 750 2         1     1     282      5        8
#> 751 2         1     1     370      5        9
#> 752 2         1     1     383      5       13
#> 753 1         1     1     412      5        9
#> 754 2         1     1     443      5       18
#> 755 1         1     1     476      5        7
#> 756 2         1     1     492      5        8
#> 757 1         1     1     600      5       17
#> 758 2         1     1     664      5        8
#> 759 2         1     1     670      5       12
#> 760 1         1     1     700      5       18
#> 761 1         1     1     721      5        8
#> 762 2         1     1     758      5        7
#> 763 1         1     1     761      5        9
#> 764 2         1     1     790      5        9
#> 765 2         1     1     791      5       10
#> 766 2         1     1     824      5       16
#> 767 2         1     1     858      5       11
#> 768 2         1     1     860      5        7
#> 769 2         1     1     861      5        8
#> 770 2         1     1     873      5        9
#> 771 1         1     1     153      6       10
#> 772 1         1     1     200      6        8
#> 773 2         1     1     285      6        9
#> 774 2         1     1     322      6        9
#> 775 2         1     1     367      6        7
#> 776 2         1     1     427      6       11
#> 777 1         1     1     454      6        7
#> 778 1         1     1     455      6        8
#> 779 2         1     1     470      6        7
#> 780 2         1     1     534      6        9
#> 781 2         1     1     539      6       17
#> 782 2         1     1     542      6        9
#> 783 1         1     1     617      6        8
#> 784 2         1     1     629      6       10
#> 785 2         1     1     695      6        8
#> 786 2         1     1     699      6       10
#> 787 2         1     1     710      6        7
#> 788 2         1     1     774      6        8
#> 789 2         1     1     811      6        9
#> 790 2         1     1     815      6        9
#> 791 2         1     1     835      6       10
#> 792 2         1     1     872      6        8
#> 793 2         1     1     298      7        7
#> 794 1         1     1     417      7       13
#> 795 2         1     1     475      7        9
#> 796 2         1     1      50      7       19
#> 797 2         1     1     808      7        8
#> 798 2         1     1     809      7        9
#> 799 1         1     1     100      8       18
#> 800 2         1     1      12      8       18
#> 801 2         1     1     141      8       20
#> 802 2         1     1     142      8       12
#> 803 2         1     1     145      8       12
#> 804 2         1     1     165      8       11
#> 805 2         1     1     184      8       10
#> 806 1         1     1     220      8        8
#> 807 2         1     1     240      8       10
#> 808 1         1     1     241      8        8
#> 809 2         1     1     242      8       10
#> 810 2         1     1     245      8        9
#> 811 2         1     1     247      8        7
#> 812 2         1     1     254      8       12
#> 813 2         1     1     292      8        7
#> 814 2         1     1     304      8       13
#> 815 2         1     1     306      8       20
#> 816 2         1     1     317      8       12
#> 817 2         1     1     321      8       11
#> 818 1         1     1     329      8        8
#> 819 2         1     1     333      8       14
#> 820 2         1     1     335      8        7
#> 821 2         1     1     346      8        9
#> 822 2         1     1     350      8        8
#> 823 1         1     1     358      8       10
#> 824 2         1     1     393      8        9
#> 825 1         1     1     415      8        9
#> 826 2         1     1     424      8       10
#> 827 2         1     1     429      8       10
#> 828 2         1     1     431      8        8
#> 829 2         1     1     446      8       11
#> 830 2         1     1     459      8        7
#> 831 2         1     1     479      8        7
#> 832 2         1     1     480      8       10
#> 833 2         1     1     484      8        9
#> 834 1         1     1     491      8       10
#> 835 1         1     1     494      8       10
#> 836 1         1     1     496      8       11
#> 837 2         1     1     523      8        9
#> 838 2         1     1     524      8       11
#> 839 1         1     1     550      8        8
#> 840 2         1     1     557      8       27
#> 841 2         1     1     558      8       25
#> 842 2         1     1     579      8        7
#> 843 1         1     1     585      8       10
#> 844 1         1     1     605      8        8
#> 845 2         1     1     616      8        9
#> 846 1         1     1     619      8        9
#> 847 2         1     1     620      8       15
#> 848 1         1     1     624      8       11
#> 849 1         1     1     637      8        8
#> 850 2         1     1     641      8       14
#> 851 2         1     1     650      8       10
#> 852 2         1     1     653      8       12
#> 853 2         1     1     655      8        7
#> 854 2         1     1     661      8       10
#> 855 1         1     1     681      8        8
#> 856 1         1     1     682      8        7
#> 857 1         1     1      71      8        7
#> 858 2         1     1     711      8        8
#> 859 1         1     1     730      8       10
#> 860 2         1     1     763      8        8
#> 861 2         1     1     765      8        7
#> 862 1         1     1     766      8        8
#> 863 2         1     1     775      8        8
#> 864 1         1     1     789      8       12
#> 865 1         1     1     795      8        7
#> 866 1         1     1     796      8       11
#> 867 2         1     1     797      8       13
#> 868 2         1     1     800      8       13
#> 869 2         1     1     801      8        9
#> 870 2         1     1     802      8       14
#> 871 2         1     1     812      8        9
#> 872 2         1     1     817      8       14
#> 873 2         1     1     820      8       10
#> 874 2         1     1     831      8       11
#> 875 2         1     1     839      8        9
#> 876 1         1     1     840      8       16
#> 877 2         1     1     841      8       14
#> 878 2         1     1     850      8        9
#> 879 2         1     1     853      8       10
#> 880 2         1     1     856      8        8
#> 881 2         1     1     857      8       10
#> 882 2         1     1     865      8       11
#> 883 2         1     1     866      8        8
#> 884 2         1     1     867      8        8
#> 885 2         1     1     868      8        9
#> 886 2         1     1     870      8        7
#> 887 2         1     1     871      8        7
#> 888 2         1     1     874      8        9
#> 889 2         1     1      89      8       25
#> 890 2         1     1     898      8        8